% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availsp.R
\name{availsp}
\alias{availsp}
\title{List of available species for ncRNA-target enrichment analysis}
\usage{
availsp()
}
\value{
An object of class data.frame
}
\description{
Prints a table with the supported species and the overall number of experimentally validated interactions by specie and ncRNA type that tienrich() uses in the statistical analysis.
}
\details{
Experimentally validated interactions between ncRNAs and their gene targets are scanned from the following databases:

- miRTarbase: miRNA-target interactions.

- RNAInter:  lncRNA, snoRNA and snRNA target interactions
}
