% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_chordMatrix.R
\name{build_chordMatrix}
\alias{build_chordMatrix}
\title{Adapt CORALIS object}
\usage{
build_chordMatrix(mti)
}
\arguments{
\item{obj}{an object from class CoralisResult or data.frame.}

\item{top}{integer indicating the top interactions to be represented in the graph. Max allowed 25 interactions.}
}
\description{
Data frame transformation of CORALIS object into chrod diagram format
}
