% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\docType{data}
\name{ids}
\alias{ids}
\title{Data of non-coding RNAs IDs}
\format{
A list with 4 slots. Each slot has a vector of non-coding RNAs IDs:
\describe{
  \item{miRNAs}{Human (Homo sapiens) miRNAs}
  \item{lncRNAs}{Human (Homo sapiens) lncRNAs}
  \item{snRNAs}{Mouse (Mus musculus) snRNAs}
  \item{snoRNAs}{Mouse (Mus musculus) snoRNAs}
}
}
\usage{
ids
}
\description{
The ids is a list object that includes 4 slots named according to different non-coding RNA types.
}
\keyword{datasets}
