% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_args}
\alias{info_args}
\title{info_input}
\usage{
info_args(a, b, c)
}
\arguments{
\item{a}{the argument 'input_list' provided to tienrich()}

\item{b}{the argument 'organism' provided to tienrich()}

\item{c}{the argument 'type' provided to tienrich()}
}
\description{
Check whether the input list in tienrich() is in valid format.
}
