% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_items}
\alias{info_items}
\title{info_items}
\usage{
info_items(k, l)
}
\arguments{
\item{k}{vector containing the ncRNAs in the input list of tienrich() for which there have been found target genes.}

\item{l}{vector containing ALL ncRNAs in the user's input list.}
}
\value{
A subset of ncRNAs IDs for which no target genes are detected.
}
\description{
Identification of target/s gene/s for the ncRNAs provided in the tienrich() function.
}
