% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_targets}
\alias{info_targets}
\title{info_targets}
\usage{
info_targets(a, b, c)
}
\arguments{
\item{a}{integer. Inherits the "min" argument from tienrich() function.}

\item{b}{character string. It inherits the "type" argument from tienrich() function.}

\item{c}{vector of target genes found for ncRNAs in the input.}
}
\value{
A vector of target genes for the ncRNAs in the input.
}
\description{
info_targets
}
\details{
Informs about the target enrichment analysis details according to the ncRNAs in the input list of tienrich.
}
