% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersummary.R
\docType{data}
\name{intersummary}
\alias{intersummary}
\title{Data of non-coding RNAs IDs}
\format{
A data frame
}
\usage{
intersummary
}
\description{
The "intersummary object" object contains the summary of interactions by specie and by type of non-codingRNA that will be used by CORALIS during the analysis.
}
\keyword{datasets}
