% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{nodeBar}
\alias{nodeBar}
\title{nodeBar}
\usage{
nodeBar(obj, top)
}
\arguments{
\item{obj}{an object from class CoralisResult or data.frame.}

\item{top}{integer indicating the top interactions to be represented in the graph.}
}
\value{
A object from class ggplot
}
\description{
Create a barplot for visualizing the number of interactions between ncRNAs and their target genes according to the ncRNA-target enrichment analysis (see ?tienrich).
}
