% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{nodeNet}
\alias{nodeNet}
\title{nodeNet}
\source{
The network graph has been built under the networkD3 package [1].
}
\usage{
nodeNet(obj, top, fixedsize)
}
\arguments{
\item{obj}{an object from class CoralisResult or data.frame.}

\item{top}{integer indicating the top interactions to be represented in the graph.}

\item{fixedsize}{Bolean. If fixedsize=True allows for interactive network, otherwise network motion is restricted.}
}
\value{
A object from class networkD3
}
\description{
Interactive Network graph representing the ncRNA-target interactions.
}
\references{
1. J.J. Allaire, Christopher Gandrud, Kenton Russell and CJ Yetman (2017). networkD3: D3 JavaScript Network Graphs from R. R package version 0.4. https://CRAN.R-project.org/package=networkD3
}
