% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{nodeVisu}
\alias{nodeVisu}
\title{nodeVisu}
\usage{
nodeVisu(obj, top = NULL, type, fixedsize = TRUE)
}
\arguments{
\item{obj}{an object from class CoralisResult or data.frame.}

\item{top}{integer indicating the top interactions to be represented in the graph.}

\item{type}{string specifying the type of visual representation. So far "network" and "barplot" are available.}
}
\description{
Visualization of ncRNA-target enrichment analysis
}
