% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiehyper}
\alias{tiehyper}
\title{The hypergeometric distribution test}
\usage{
tiehyper(
 id,
 m,
 n,
 k,
 x,
 nhits
)
}
\arguments{
\item{id}{character string indicating the name of the  target gene.}

\item{m}{integer. The number of ncRNAs that interact with gene "id".}

\item{n}{integer. The number of ncRNAs that do not interact with gene "id".}

\item{k}{integer. The number of ncRNAs in the input of tienrich (see ?tienrich) that interact with gene "id".}

\item{x}{vector representing the range of interactions from 0 to min(k, m).}

\item{nhits}{integer. Number of interactions between gene "id" and ncRNAs in the input of tienrich().}
}
\value{
A object from class data.frame
}
\description{
Statistical source function for target enrichment analysis. The hypergeometric distribution is used to estimate the probability of having k interactions between the ncRNAs in the input of  tienrich and the target gene according to the source databases (see ?tienrich()).
}
\details{
This function employs the hypergeometric distribution approach to compute the probability distribution of the interactions between the ncRNAs and their target genes (see ?dhyper for detailed info)[1].
Note that this function also provides the odds ratio (OR), the relative risk (RR) as well as the standard error and their confidence intervals.
}
\examples{
#Define the arguments
id <- "BCL2" #Gene id
m  <- 38
n  <- 2561
N  <- n + m
k  <- 4
nhits <- 2
x  <- c(0:min(k,m))
#hypergeomentric test.
tiehyper(id = id, m = m, n = n, k = k, x = x, nhits = nhits)
}
\references{
1. R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
}
\seealso{
?tienrich and ?dhyper
}
