% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetInteraction_enrichment.R
\name{tienrich}
\alias{tienrich}
\title{non-coding RNA target enrichment analysis}
\usage{
tienrich(
 input_list,
 organism,
 type
 min = 1,
 fdr = 1
)
}
\arguments{
\item{input_list}{a vector of ncRNAs IDs in miRBase format for microRNAs (ie: 'hsa-miR-3196') or Official Gene Symbol format for lncRNAs, snoRNAs or snRNAs (i.e: 'RUNX2').}

\item{type}{character string indicating the desired type of ncRNA target enrichment analysis (ie:'miRNA_mRNA' for microRNA-target interactions or 'lncRNA_mRNA' for lncRNA-target interactions, 'snoRNA_mRNA' for snoRNA-target interactions or 'snRNA_mRNA' for snRNA-target interactions).}

\item{organism}{character string specifying the organism name (i.e: "Homo sapiens"). Run availsp() to list supported organisms and a summary of interactions by specie and ncRNA type.}

\item{min}{integer specifying the minimum number of interactions that target genes must have with the ncRNAs in the input.}

\item{fdr}{adjusted p-value (Benjamin & Hochberg approach) threshold at which enriched target-genes are found.}
}
\value{
An object of class CoralisResult.
}
\description{
Conducts a target enrichment analysis for a set of ncRNAs. Only experimentally validated interactions between ncRNAs and their target genes are considered (see Details section).
}
\details{
The tienrich function analyses the interactions between ncRNAs and their target genes thought enrichment/overrepresentation analysis.
To this end, this function employs the hypergeometric distribution test to estimate the p-values of the interaction between the ncRNAs provided in the input_list and their target genes.
So far, only experimentally validated interactions listed in miRTarbase (for microRNAs-target interactions)[1] and RNAInter Database (for lncRNAs, snoRNAs or snRNAs target interactions)[2] are used as background in the enrichment analysis.
Additionally, tienrich provides other interesting statistical parameters such the odds ratio (OR), the conficence intevalsand the adjusted p-value (false discovery rate -FDR- using Benjamin-Hochberg correction).
}
\examples{
# vector of ncRNA IDs
data(rnasID)
head(ids)
# Homo sapiens -----> ids[['miRNAs']]
# Homo sapiens -----> ids[['lncRNAs']]
# Mus musculus -----> ids[['snRNAs']]
# Mus musculus -----> ids[['snoRNAs']]

# Run target enrichment analysis:
tar_mir <- tienrich(input_list = ids[['miRNAs']],  min = 2, fdr = 1, organism = 'Homo sapiens', type = 'miRNA_mRNA')
tar_lnc <- tienrich(input_list = ids[['lncRNAs']], min = 2, fdr = 1, organism = 'Homo sapiens', type = 'lncRNA_mRNA')
tar_sn  <- tienrich(input_list = ids[['snRNAs']],  min = 2, fdr = 1, organism = 'Mus musculus', type = 'snRNA_mRNA')
tar_sno <- tienrich(input_list = ids[['snoRNAs']], min = 2, fdr = 1, organism = 'Mus musculus', type = 'snoRNA_mRNA')
}
\references{
1. Huang HY et al. 2020. miRTarBase 2020: updates to the experimentally validated microRNA-target interaction database. Nucleic Acids Res. 2020 Jan 8;48(D1):D148-D154. doi: 10.1093/nar/gkz896. PMID: 31647101; PMCID: PMC7145596.

2. Lin Y et al. 2020. RNAInter in 2020: RNA interactome repository with increased coverage and annotation. Nucleic Acids Res. 2020 Jan 8;48(D1):D189-D197. doi: 10.1093/nar/gkz804. PMID: 31906603; PMCID: PMC6943043.
}
\seealso{
CORALIS::tiehyper()
}
\author{
Daniel Valle-Millares
}
