% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{nodeVisu}
\alias{nodeVisu}
\title{nodeVisu}
\usage{
nodeVisu(obj, top = NULL, type, fixedsize = TRUE)
}
\arguments{
\item{obj}{an object from class CoralisResult or data.frame.}

\item{top}{integer indicating the top interactions to be represented in the graph.}

\item{type}{string specifying the type of visual representation. So far "network" and "barplot" are available.}

\item{fixedsize}{Bolean. If fixedsize=True allows for interactive network, otherwise network motion is restricted.}
}
\description{
Visualization of ncRNA-target enrichment analysis.
}
\examples{
# vector of ncRNA IDs
data(rnasID)
head(ids)
# Target enrichment analysis:
tar_mir <- tienrich(input_list = ids[['miRNAs']],
                    min = 2, fdr = 1, organism = 'Homo sapiens', type = 'miRNA_mRNA')
# To visualize the target enrichment analysis as barplot.
bar <- nodeVisu(obj = tar_mir, top = 25, type = "barplot")
print(bar)
# To visualize the target enrichment analysis as network graph.
net <- nodeVisu(obj = tar_mir, top = 25, type = "network")
print(net)
}
